Move OS-specific logic into separate files
While we don't officially support non-Linux/Darwin builds
we should still keep the OS-specific logic in separate files
and follow best practices.

Fixes #780.
